rename ioc to cmds
