+++ src/ec_cli/autocomplete.py
@@ -11,9 +11,9 @@ import typer
 
 import ec_cli.globals as globals
 import ec_cli.shell as shell
+from ec_cli.cmds.k8s_commands import check_namespace
 from ec_cli.docker import Docker
 from ec_cli.git import create_ioc_graph
-from ec_cli.ioc.k8s_commands import check_namespace
 from ec_cli.logging import log
 from ec_cli.utils import cleanup_temp
