+++ src/ec_cli/cmds/kubectl.py
