+++ tests/data/iocs/bl45p-ea-ioc-01/config/ioc.yaml
@@ -1,4 +1,4 @@
-# yaml-language-server: $schema=https://github.com/epics-containers/ioc-adaravis/releases/download/23.9.4/ibek.ioc.schema.json
+# yaml-language-server: $schema=https://github.com/epics-containers/ioc-adaravis/releases/download/23.9.4/ibek.cmds.schema.json
 
 ioc_name: bl45p-ea-ioc-01
 description: Sample and overview cameras IOC for BL45P
